+++ src/components/index.js
@@ -19,7 +19,6 @@ export { default as Container } from './Container';
 
 /** Still needs to be refactored into functional */
 export { default as MTableBody } from './m-table-body';
-export { default as MTableBodyRow } from './m-table-body-row';
 /** HAVING ISSUES WITH THE REFACTORED VERSIONS OF: */
 export { default as MTableEditField } from './m-table-edit-field';
 export { default as MTableEditCell } from './m-table-edit-cell';
@@ -32,7 +31,9 @@ export { default as MTableEditCell } from './m-table-edit-cell';
 // Trying to keep these in alphabetical order
 export { default as MTableAction } from './MTableAction';
 export { default as MTableActions } from './MTableActions';
+export { default as MTableBodyRow } from './MTableBodyRow';
 export { default as MTableCell } from './MTableCell';
+export { default as MTableCustomIcon } from './MTableCustomIcon';
 export { default as MTableEditRow } from './MTableEditRow';
 export { default as MTableFilterRow } from './MTableFilterRow';
 export { default as MTableGroupbar } from './MTableGroupbar';